Eon planner

+++ serial/SerialCore/Loop.cpp
@@ -253,7 +253,7 @@ String Loop::GetTreeString(int indent) {
 	String pre;
 	pre.Cat('\t', indent);
 	
-	s << ".." << (name.IsEmpty() ? "unnamed" : "\"" + name + "\"") << "[" << id << "]\n";
+	s << ".." << (name.IsEmpty() ? "unnamed" : "\"" + name + "\"") << "[" << (int)id << "]\n";
 	
 	for (AtomBaseRef& a : atoms)
 		s << a->ToString();
